pyproject.toml
[project]
name = "captcha-fast-detector"
version = "0.2.0"
description = "Step 1: ROI Calibrator for consistent-layout captchas"
authors = [
    {name = "Captcha Fast Detector Team"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "pillow>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.0.0",
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["captcha_detector"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
